refactor: Moved model and entity classes to different sub packages.
JPA entity and MongoDB models moved to new sub packages.
Additional sub package '*.data.dto" created for the new DTO classes
used in the controller to implement the load requirement:

  " The project uses its own JSON representation for REST instead of
serializing the entities directly.
