+++ src/main/java/com/udacity/course3/reviews/data/model/Review.java
@@ -1,8 +1,12 @@
-package com.udacity.course3.reviews.model;
+package com.udacity.course3.reviews.data.model;
 
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+
+import javax.validation.constraints.Max;
+import javax.validation.constraints.Min;
+import javax.validation.constraints.NotNull;
 
 import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
@@ -20,13 +24,20 @@ public class Review {
   @Id
   private ObjectId  id;
   
+  @NotNull
   private String    title;
   
+  @NotNull
   private String    body;
   
+  @NotNull
   private String    author;
   
   private Date      createdTime;
+  
+  @Min(value = 1)
+  @Max(value = 5)
+  private Integer   rating;    
   
   // :INFO: This is the primary key of the mysql database. 
   @JsonIgnore
@@ -100,6 +111,16 @@ public class Review {
 
   public void setProductId(Integer product_id) {
     this.productId = product_id;
+  }
+
+
+  public Integer getRating() {
+    return rating;
+  }
+
+
+  public void setRating(Integer rating) {
+    this.rating = rating;
   }
 
 }
