+++ src/main/java/com/udacity/course3/reviews/repository/ProductRepository.java
@@ -2,7 +2,7 @@ package com.udacity.course3.reviews.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.udacity.course3.reviews.entity.Product;
+import com.udacity.course3.reviews.data.entity.Product;
 
 public interface ProductRepository extends JpaRepository<Product, Integer> {
