+++ src/main/java/com/udacity/course3/reviews/repository/ReviewRepository.java
@@ -7,7 +7,7 @@ import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import com.udacity.course3.reviews.model.Review;
+import com.udacity.course3.reviews.data.model.Review;
 
 @Repository
 public interface ReviewRepository extends MongoRepository<Review, ObjectId> {
