+++ src/test/java/com/udacity/course3/reviews/ReviewControllerTest.java
@@ -25,7 +25,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
-import com.udacity.course3.reviews.model.Review;
+import com.udacity.course3.reviews.data.model.Review;
 import com.udacity.course3.reviews.service.ReviewService;
 import com.udacity.course3.reviews.util.DummyDataUtil;
